Add CRUD operations to RegistrationController.

+++ app/Http/Controllers/RegistrationController.php
@@ -2,6 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use Carbon\Carbon;
+use App\Meeting;
+use App\User;
 use Illuminate\Http\Request;
 
 class RegistrationController extends Controller
@@ -22,26 +25,31 @@ class RegistrationController extends Controller
         $meeting_id = $request->input('meeting_id');
         $user_id = $request->input('user_id');
         
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
+        $meeting = Meeting::findOrFail($meeting_id);
+        $user = User::findOrFail($user_id);
+
+        $message = [
+            'msg' => 'User is already registered for meeting',
+            'user' => $user,
+            'meeting' => $meeting,
+            'unregister' => [
+                'href' => 'api/v1/meeting/registration/' . $meeting->id,
+                'method' => 'DELETE'
             ]
         ];
 
-        $user = [
-            'name' => 'Name'
-        ];
+        if($meeting->users()->where('users.id', $user->id)->first()){
+            return response()->json($message, 404);
+        }
+
+        $user->meetings()->attach($meeting);
 
         $response = [
             'msg' => 'User registered for meeting.',
             'meeting' => $meeting,
             'user' => $user,
             'unregister' => [
-                'href' => 'api/v1/meeting/registration/1',
+                'href' => 'api/v1/meeting/registration/'. $meeting->id,
                 'method' => 'DELETE'
             ]
         ];
@@ -57,26 +65,15 @@ class RegistrationController extends Controller
     public function destroy($id)
     {
 
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
-
-        $user = [
-            'name' => 'Name'
-        ];
+        $meeting = Meeting::findOrFail($id);
+        $meeting->users()->detach();
 
         $response = [
             'msg' => 'User unregistered for meeting.',
             'meeting' => $meeting,
-            'user' => $user,
+            'user' => 'tbd',
             'register' => [
-                'href' => 'api/v1/meeting/registration/1',
+                'href' => 'api/v1/meeting/registration',
                 'method' => 'POST',
                 'params' => 'user_id, meeting_id'
             ]
